ipc: deduplicate some unix-specific code
Cleans up and splits out UAPIOpen to its own file.

[zx2c4: changed const to var for socketDirectory]
